Fix plotClonalFrequency on genome and region plots
As told in issue #12:

 In the code for the function, clusters expands to

```r
> clusters
   ClonalCluster CellularPrevalence
1:             2          0.7714099
2:             3          0.6370672
3:             1          1.0000000
4:             4          0.5077241
```

When plotting, the code does

```r
if (nrow(clusters) > 0) {
            for (j in 1:length(clusters[, 1])) {
```

However:

```r
> 1:length(clusters[, 1])
[1] 1
```

This PR fixes it by iterating on the actual data.table instead of using
`length()`. Notice that the diff may show a much larger change than
expected because I also fixed part of the indentation.

This fixes issue #12.